Moved studentDb DI to pandautils.
