+++ apps/student/src/main/java/com/instructure/student/di/DatabaseModule.kt
@@ -6,12 +6,6 @@ import com.instructure.pandautils.room.appdatabase.AppDatabase
 import com.instructure.pandautils.room.appdatabase.appDatabaseMigrations
 import com.instructure.pandautils.room.calendar.CalendarFilterDatabase
 import com.instructure.pandautils.room.calendar.calendarDatabaseMigrations
-import com.instructure.pandautils.room.studentdb.StudentDb
-import com.instructure.pandautils.room.studentdb.entities.daos.CreateFileSubmissionDao
-import com.instructure.pandautils.room.studentdb.entities.daos.CreatePendingSubmissionCommentDao
-import com.instructure.pandautils.room.studentdb.entities.daos.CreateSubmissionCommentFileDao
-import com.instructure.pandautils.room.studentdb.entities.daos.CreateSubmissionDao
-import com.instructure.pandautils.room.studentdb.studentDbMigrations
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -41,38 +35,5 @@ class DatabaseModule {
         ) // We need to have the same db name as in the Flutter calendar
             .addMigrations(*calendarDatabaseMigrations)
             .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideStudentDb(@ApplicationContext context: Context): StudentDb {
-        return Room.databaseBuilder(context, StudentDb::class.java, "student.db")
-            .addMigrations(*studentDbMigrations)
-            .fallbackToDestructiveMigration()
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideCreateSubmissionDao(studentDb: StudentDb): CreateSubmissionDao {
-        return studentDb.submissionDao()
-    }
-
-    @Provides
-    @Singleton
-    fun provideCreateFileSubmissionDao(studentDb: StudentDb): CreateFileSubmissionDao {
-        return studentDb.fileSubmissionDao()
-    }
-
-    @Provides
-    @Singleton
-    fun provideCreatePendingSubmissionCommentDao(studentDb: StudentDb): CreatePendingSubmissionCommentDao {
-        return studentDb.pendingSubmissionCommentDao()
-    }
-
-    @Provides
-    @Singleton
-    fun provideCreateSubmissionCommentFileDao(studentDb: StudentDb): CreateSubmissionCommentFileDao {
-        return studentDb.submissionCommentFileDao()
     }
 }
